+++ src/lib.rs
@@ -1,6 +1,6 @@
 pub mod calculate_mean;
+pub mod errors;
 pub mod goals;
+pub mod grades;
 pub mod saving;
 pub mod subject;
-pub mod grades;
-pub mod errors;