+++ CHANGELOG.md
@@ -1,14 +1,19 @@
 # v0.4.0 (05/31/2026)
 
-The backend of Scoreledger has been partially rebuilt.
+The backend of Scoreledger has been partially rebuilt. This is one major step to reaching stability.
 
 ## Added
 - New error messages, they're easier to read and more precise.
+- Removing a subject also removes the grade related to it.
 
 ## Changed
 - The backend of Scoreledger has been partially rebuilt. This is to ensure that functions related to working with subjects, goals, grades, saves, etc are dedicated to only that, leaving everything with logging to be handled by the CLI.
 - Made the code more idiomatic. Less process::exit functions, use Result instead.
 - Logic related to handling grades moved to it's own file (it was formerly in the subjects file.)
+
+## Bug Fixes
+- "Delete a subject" option no longer panics when there are no subjects to delete.
+- "Delete a goal" option no longer panics when there are no goals to delete.
 
 # v0.3.1 (05/30/2026)